Rename "gws" to "grid"; replace "gws" with "grid-lite".
