+++ lib/grid/basic.lua
@@ -1,6 +1,6 @@
 -- basic information
 -- by Qige
--- 2017.01.05
+-- 2017.01.05/2017.01.16
 
 basic = {}
