+++ lib/grid/grid-lite.conf
@@ -1,9 +1,9 @@
 # 6Harmonics Qige
-# 2017.02.08
+# 2017.02.16
 
-config gws_web 'v3'
+config grid-lite 'v1'
 	option user 'admin'
 	option pass 'admin'
 	#option root '192.168.1.211'
-	option version '3.1.020817'
+	option version '1.0-1-160217'
