+++ lib/grid/kpi.lua
@@ -1,11 +1,11 @@
 -- kpi: gws/network/baseband/stas nw
 -- by Qige
--- 2017.01.05
+-- 2017.01.05/2017.01.16
 
-require 'wui.base.cgi'
-require 'wui.bb'
-require 'wui.gws'
-require 'wui.nw'
+require 'grid.base.cgi'
+require 'grid.bb'
+require 'grid.gws'
+require 'grid.nw'
 
 
 kpi = {}
