+++ www/grid/index.html
@@ -1,13 +1,9 @@
 <!DOCTYPE html>
 <html>
 <head>
-	<!--Import Google Icon Font-->
 	<link href="res/google/icon.css" rel="stylesheet">
-	<!--Import materialize.css-->
 	<link type="text/css" rel="stylesheet" href="res/materialize/css/materialize.min.css"  media="screen,projection" />
 	<link type="text/css" rel="stylesheet" href="res/app.css" media="screen,projection" />
-
-	<!--Let browser know website is optimized for mobile-->
 	<meta name="viewport" content="width=device-width, initial-scale=1.0" />
 	<title>GRID Lite from 6Harmonics</title>
 </head>
@@ -89,10 +85,9 @@
 		</div>
 	</div>
 </footer>
-
-<!--Import jQuery before materialize.js-->
 <script type="text/javascript" src="res/jquery.min.js"></script>
 <script type="text/javascript" src="res/materialize/js/materialize.min.js"></script>
 <script type="text/javascript" src="res/login.js"></script>
+<!-- Design & Powered by Qige from 6Harmonics/6WiLink -->
 </body>
 </html>
