+++ www/index.html
@@ -1,9 +1,10 @@
 <!DOCTYPE html>
 <html>
 <head>
-<meta http-equiv="refresh" content="0; /gws/" />
+	<meta http-equiv="refresh" content="0; /grid/" />
+	<title>Redirecting to GRID Lite ...</title>
 </head>
 <body>
-Redirecting to GWS WebUI ...
+	Redirecting to GRID Lite ...
 </body>
 </html>
